🔨 Convert '__LINE__' to 'size_t'
The ALLOY_ASSERT macro internally uses unsigned values for line
numbers, yet __LINE__ is signed. This leads to some compilers
complaining about implicit conversions when being extremely
pedantic.

In general, it is safe to assume that __LINE__ will never be negative,
and as a result is simply static_cast-ed to size_t

+++ lib/alloy-core/include/alloy/core/assert.hpp
@@ -74,7 +74,7 @@
           "assertion failure: condition '" \
           ALLOY_STRINGIZE(condition) \
           "' failed.", \
-          file_name, line_number, function_name \
+          file_name, static_cast<std::size_t>(line_number), function_name \
         ); \
       }( __FILE__, __LINE__, __func__ ) )
 
@@ -90,7 +90,7 @@
           ALLOY_STRINGIZE(condition) \
           "' failed with message \"" \
           message "\"", \
-          file_name, line_number, function_name \
+          file_name, static_cast<std::size_t>(line_number), function_name \
         ); \
       }( __FILE__, __LINE__, __func__ ) )
